Extract pin generation to own function
